+++ tests/test_timing_tracker.py
@@ -26,9 +26,9 @@ class TestTimingTrackerBehavior:
         clock = FakeClock()
         tt = TestTimingTracker(clock=clock)
 
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="test_one"'))
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="test_one"'))
         clock.advance(2.5)
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="test_two"'))
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="test_two"'))
         clock.advance(1.0)
         tt.finalize()
 
@@ -39,9 +39,9 @@ class TestTimingTrackerBehavior:
         clock = FakeClock()
         tt = TestTimingTracker(slow_threshold=5.0, clock=clock)
 
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="fast"'))
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="fast"'))
         clock.advance(1.0)
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="slow"'))
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="slow"'))
         clock.advance(10.0)
         tt.finalize()
 
@@ -52,7 +52,7 @@ class TestTimingTrackerBehavior:
     def test_report_empty_when_no_slow(self):
         clock = FakeClock()
         tt = TestTimingTracker(slow_threshold=5.0, clock=clock)
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="fast"'))
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="fast"'))
         clock.advance(1.0)
         tt.finalize()
         assert tt.report() == ""
@@ -60,7 +60,7 @@ class TestTimingTrackerBehavior:
     def test_report_formats_slow_tests(self):
         clock = FakeClock()
         tt = TestTimingTracker(slow_threshold=5.0, clock=clock)
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="slow"'))
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="slow"'))
         clock.advance(8.0)
         tt.finalize()
 
@@ -72,7 +72,7 @@ class TestTimingTrackerBehavior:
     def test_test_start_with_timeout_annotation(self):
         clock = FakeClock()
         tt = TestTimingTracker(clock=clock)
-        tt.feed(_crc('ETST:TEST:START suite="GPS" name="Nav rate" timeout=30'))
+        tt.feed(_crc('ETST:CASE:START suite="GPS" name="Nav rate" timeout=30'))
         clock.advance(3.0)
         tt.finalize()
         assert tt.durations["GPS/Nav rate"] == 3.0
@@ -86,7 +86,7 @@ class TestTimingTrackerBehavior:
     def test_bytes_input(self):
         clock = FakeClock()
         tt = TestTimingTracker(clock=clock)
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="test"').encode())
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="test"').encode())
         clock.advance(1.0)
         tt.finalize()
         assert "Suite/test" in tt.durations
@@ -94,7 +94,7 @@ class TestTimingTrackerBehavior:
     def test_reset_clears_all(self):
         clock = FakeClock()
         tt = TestTimingTracker(clock=clock)
-        tt.feed(_crc('ETST:TEST:START suite="Suite" name="test"'))
+        tt.feed(_crc('ETST:CASE:START suite="Suite" name="test"'))
         clock.advance(1.0)
         tt.finalize()
         tt.reset()